Revert split of rollup plugin
